Add files via upload
The National Flag of the Lao People's Democratic Republic is dark blue with red edges and a white moon in the middle of the flag. The height of the flag is two- thirds of its width. The height of each red edge is one-half of the height of the dark blue area . It is using only pure HTML & CSS with the help of CSS positioning properties .